requirements.txt
nextcord
requests
Flask
python-dotenv